descw-1280 modify backend utilities for cleanliness
